Continued Localization
Partly added some of the remaining localization I need to continue
getting to. It's nowhere near done.

Special thanks to Iustinius for his additional names for the deity list
of the Kong Faith, and to Empoleon Blownaparte for his amazing
description for the religion he conceived, the Temple of Form!