Species Trait Implementation I
It's time for the real hard stuff. I am writing all the Traits
Avrolancaster made into the game before adding missing ones (e.g.
Tanoombas, Twilighters).

After this will come filler "cultures" that use GFX for each species to
ensure that Gloombas and the like can show up in the game. Events will
have to be made to flip their culture to an in-game one.