Updated the Files with the Correct Traits (Also there is something I need to tell TurtleShroom about the Jellyfish Isle)
I took noticed that in the Jellyfish Island character file, the only character in there uses a Latin name, Arvolancaster may have intended for the Jellyfish Island cultural nameset to be Latin instead of Portuguese. The current Jellyfish Island cultural nameset can be given to the Eel Island which is the only set of islands of the Delfino Islands that don't have a cultural nameset.