Updated All of the Namesets For the Delfino Group
Each nameset is based on a Romance language. Delfinite is still Italian, fish_islands is still Spanish,  crab_island is French, turtle_island is now a reconstruction of African Latin, eel_island is still Portuguese, Jellyfishing is still Latin, noki_bay is now Occitan, pianta is now Romanian, and noki is now Old Frankish, only non-Romance language in the group, my reasoning is I'm assuming these Noki are gonna be separated from the rest of the Noki of the Sunshine Isles which would be fitting as this language is uninfluenced by the languages of the Sunshine Isles.

At some point down the road, I'm gonna expand the namelist with even more names.