Species Implementation XI
As promised, I chose to work on the Mod today (the promised day after
Thanksgiving that I would serve) and managed to get a lot of species
localization done!

I would have gotten more, but my sister is home from college and she
wants to play "Mario Kart" with me.